Rename: 폴더명 변경(node -> server)
