+++ server/routes/manage.js
